LP-CR-679 extract benchmark workflow service
