+++ quality/refactor_health_report.md
@@ -1,7 +1,7 @@
 # Lotus Performance Refactor Health Report
 
 Report date: 2026-06-04
-Branch: `feat/performance-hardening-wave-11`
+Branch: `feat/performance-hardening-wave-12`
 Baseline source: `quality/baseline_report.md`
 Report mode: phase-zero scorecard; no blocking gate is introduced by this artifact.
 
@@ -24,9 +24,9 @@ link the commit, command, or CI artifact that proves the change.
 
 | Metric | Baseline | Current | Status | Evidence |
 | --- | ---: | ---: | --- | --- |
-| Python files | 480 | 480 | measured | `rg --files -g '*.py'` |
+| Python files | 480 | 531 | measured | `rg --files -g '*.py'` |
 | Python package markers | 18 | 18 | measured | recursive `__init__.py` count |
-| Python LOC | 104,454 | 104,454 | measured | recursive `.py` line count |
+| Python LOC | 104,454 | 109,595 | measured | recursive `.py` line count |
 | Largest Python file LOC | 2,399 | 2,399 | measured | largest-file inventory in baseline report |
 | Largest production file LOC | 1,156 | 1,156 | measured | `app/services/lineage_metadata_store.py` |
 | Duplicate code hotspots | 0 | 0 | measured | `quality/duplicate_code_inventory.md` via `scripts/python_duplicate_code_inventory.py` with `--min-lines 12` |
@@ -50,7 +50,7 @@ link the commit, command, or CI artifact that proves the change.
 | Routers importing infrastructure directly | unknown | 0 | measured | `ROUTER_DIRECT_BOUNDARY_IMPORT` in `quality/architecture_boundary_inventory.md` |
 | Domain/application importing framework or infra code | unknown | 0 | measured | `DOMAIN_INFRA_OR_FRAMEWORK_IMPORT` absent from `quality/architecture_boundary_inventory.md` |
 | Large production service hotspots | 3 | 3 | measured | `lineage_metadata_store.py`, `compute_job_store.py`, `stateful_input_service.py` exceed 1,000 LOC |
-| Router/middleware oversized function findings (`--threshold 80`) | unknown | 4 | measured | `quality/router_middleware_thinness_inventory.md` via `scripts/python_router_middleware_thinness_inventory.py` |
+| Router/middleware oversized function findings (`--threshold 80`) | unknown | 3 | measured | `quality/router_middleware_thinness_inventory.md` via `scripts/python_router_middleware_thinness_inventory.py` |
 
 ## API Quality
 
@@ -68,8 +68,8 @@ link the commit, command, or CI artifact that proves the change.
 
 | Metric | Baseline | Current | Status | Evidence |
 | --- | ---: | ---: | --- | --- |
-| Test modules | 228 | 228 | measured | `rg --files tests -g 'test_*.py'` |
-| Collected tests | 2,035 | 2,035 | measured | `python -m pytest --collect-only -q` |
+| Test modules | 228 | 251 | measured | `rg --files tests -g 'test_*.py'` |
+| Collected tests | 2,035 | 2,117 | measured | `python -m pytest --collect-only -q` |
 | Line coverage | unknown | unknown | not-yet-measured | coverage run not captured in baseline slice |
 | Branch coverage | unknown | unknown | not-yet-measured | branch coverage not configured as a scorecard input |
 | Integration/API/runtime test functions | unknown | 453 | measured | `quality/test_taxonomy_inventory.md` via `scripts/python_test_taxonomy_inventory.py` |
