+++ quality/router_middleware_thinness_inventory.md
@@ -2,15 +2,14 @@
 
 | Metric | Value |
 | --- | ---: |
-| Router and middleware oversized function findings | 4 |
-| Oversized router functions | 4 |
+| Router and middleware oversized function findings | 3 |
+| Oversized router functions | 3 |
 | Oversized middleware functions | 0 |
 
 ## Findings
 
 | Rank | Kind | File | Function | Lines |
 | ---: | --- | --- | --- | ---: |
-| 1 | router | `app/api/endpoints/benchmark.py:63` | `calculate_benchmark_endpoint` | 149 |
-| 2 | router | `app/api/endpoints/contribution.py:129` | `calculate_contribution_endpoint` | 127 |
-| 3 | router | `app/api/endpoints/performance.py:349` | `calculate_attribution_endpoint` | 108 |
-| 4 | router | `app/api/endpoints/returns_series.py:118` | `get_returns_series` | 92 |
+| 1 | router | `app/api/endpoints/contribution.py:129` | `calculate_contribution_endpoint` | 127 |
+| 2 | router | `app/api/endpoints/performance.py:349` | `calculate_attribution_endpoint` | 108 |
+| 3 | router | `app/api/endpoints/returns_series.py:118` | `get_returns_series` | 92 |
